test edit authorization

+++ Backend/studycave-rest-service/src/main/java/studycave/application/SetController.java
@@ -4,11 +4,16 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 import java.util.concurrent.ThreadLocalRandom;
 
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,6 +21,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -233,8 +239,20 @@ public class SetController {
 	}
 
 	@DeleteMapping("/{id}")
-	public void deleteSet(@PathVariable(required = true) Long id) {
-		setRepository.deleteById(id);
+	public ResponseEntity deleteSet(@RequestHeader(value = "Authorization",required=false) String headerStr,@PathVariable(required = true) Long id) {
+		
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		String currentPrincipalName = authentication.getName();
+		Long userId = userRepository.findByUsername(currentPrincipalName).get().getId();
+
+		Optional<Set> set = setRepository.findById(id);
+		if (set.isPresent()) {
+			if (userId.equals(set.get().getIdOwner())) {
+				setRepository.deleteById(id);
+				return new ResponseEntity(HttpStatus.OK);
+			}
+		}
+		return new ResponseEntity("Access Forbidden", HttpStatus.FORBIDDEN);
 	}
 
 	@GetMapping
